requirements.txt
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
flair==0.10
Flask==2.2.1
nltk==3.6.2
numpy==1.21.6
openai==0.16.0
pandas==1.3.5
python-Levenshtein==0.12.2
rake-nltk==1.0.6
requests==2.28.1
roman==3.3
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
tqdm==4.62.3
transformers==4.21.2